Docker-files/requirements.txt
faster-whisper
piper-tts
huggingface_hub
numpy<2
pyaudio
sounddevice
soundfile
fastmcp
fastapi
tflite-runtime
httpx
mcp
openai
requests
openwakeword
uvicorn
python-multipart
flask
pyyaml
tqdm
